Fix: Calibration now expects data to be provided as ACES2065-1 in the core, we now get the input plate gamuts from the project ocio config, and we apply the conversion to ACES2065-1 in the sampling, we then pass ACES2065-1 samples to the calibration this removes the need for the user to convert from camera native encodings to ACES2065-1 which seems to cause 90% of the issues for users.

+++ src/open_vp_cal/imaging/imaging_utils.py
@@ -964,7 +964,7 @@ def calculate_white_balance_matrix_from_img_buf(
     red_mult_val = average_rgb[1] / average_rgb[0]
     blue_mult_val = average_rgb[1] / average_rgb[2]
     white_balance_matrix = np.asarray(
-        [[red_mult_val, 0.0, 0.0], [0.0, 1.0, 0.0], [0.0, 0.0, blue_mult_val]])
+        [[red_mult_val, 0.0, 0.0], [0.0, 1.0, 0.0], [0.0, 0.0, blue_mult_val]], dtype=np.float32)
     return white_balance_matrix
